HACK: Change ordering of protocol events so clients see tablet burst
We don't see the initial burst of tablet events when connecting a new client.
Based on my poking around, it looks like this is because the order of events
is incorrect. When a client (or device) appears, the seat has the Tablet
create a new resource specific to the pair. Prior to returning the resource
to the seat, the Tablet sends all the device details. After this occurs,
the resource is returned and the seat sends a notification to the client
about the tablet. Changing the order so that the burst is sent after the
notification is sent apparently fixes this.

I'm guessing that the client processes events in order, albeit asynchronously.
This would imply that the events were dropped since the client had not
registered a listener at the time the events were recieved. By changing the
order, we register a listener as part of the first event, ensuring that we
see the subsuquent ones as well.

Although this explanation makes sense, my implementation is a bit of a hack.
Based on how the Tool works, the Tablet should be in charge of sending the
burst (like it had been). An alternative fix would be to have the Tablet
send the seat notification, but again, this should be (and was) the responsibility
of the Seat. The Tool code should be studied more closely so that we can
replicate its pattern in Tablet.