HACK? Rely on client to free tablet and tool
The tablet protocol specifies that the 'destroy' methods on the tablet
and tool should be called in response to the 'removed' signal. Mutter
destroys each resource after sending the signal, however, which I think
might be the cause of the crashes that I'm seeing when disconnecting a
tablet in Xwayland. I'm not convinced its the case, however, since the
crashes occur even if I don't destroy the objects within the client
(meaning that I shouldn't be working with an already-dead object...)